feat(roles): replace manage-roles dialog with dedicated settings page
Removes the 2000-line manage-roles-dialog and replaces it with a
standalone /settings/roles page and org-role-detail view, giving roles
a proper place in the settings nav instead of a floating dialog.
