+++ apps/mesh/src/web/components/tool-set-selector.tsx
@@ -22,6 +22,8 @@ export interface ToolSetSelectorProps {
   onToolSetChange: (toolSet: Record<string, string[]>) => void;
   /** Virtual MCP ID to exclude from selection (prevents self-reference) */
   excludeVirtualMcpId?: string;
+  /** Controlled search query — if provided, hides the internal search bar */
+  searchQuery?: string;
 }
 
 interface ConnectionItemProps {
@@ -140,8 +142,13 @@ export function ToolSetSelector({
   toolSet,
   onToolSetChange,
   excludeVirtualMcpId,
+  searchQuery: controlledSearchQuery,
 }: ToolSetSelectorProps) {
-  const [searchQuery, setSearchQuery] = useState("");
+  const [internalSearchQuery, setInternalSearchQuery] = useState("");
+  const searchQuery =
+    controlledSearchQuery !== undefined
+      ? controlledSearchQuery
+      : internalSearchQuery;
   const deferredSearchQuery = useDeferredValue(searchQuery);
   const [filterMode, setFilterMode] = useState<FilterMode>("all");
   const [showMobileDetail, setShowMobileDetail] = useState(false);
@@ -358,12 +365,14 @@ export function ToolSetSelector({
           showMobileDetail && "hidden md:flex",
         )}
       >
-        {/* Search Input */}
-        <CollectionSearch
-          value={searchQuery}
-          onChange={setSearchQuery}
-          placeholder="Search MCP Servers..."
-        />
+        {/* Search Input — hidden when controlled from outside */}
+        {controlledSearchQuery === undefined && (
+          <CollectionSearch
+            value={internalSearchQuery}
+            onChange={setInternalSearchQuery}
+            placeholder="Search MCP Servers..."
+          />
+        )}
 
         {/* Filter Buttons */}
         <div className="flex gap-1 p-2 border-b border-border">
